pyproject.toml
[project]
name = "my-hermetic-agent"
version = "0.1.0"
description = "🚧 WIP—Tinkering with embeddings, agent orchestration patterns, local LLM behavior, and instruction-layer experimentation. Expect breakthroughs, breakages, and the occasional accidental summon."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ollama>=0.4.0",
    "pyyaml>=6.0.1",
    "psycopg2-binary>=2.9.9",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "msgspec>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mdformat>=0.7.17",
    "mdformat-gfm>=0.3.5",
    "mdformat-frontmatter>=2.0.0",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
addopts = "-v"
